Refactored main screen folder structure
