+++ public/app/screens/main/navigation/style.scss
