+++ public/app/screens/main/sections/devices/device/device.jsx
