+++ public/app/screens/main/sections/devices/device/index.js
