+++ public/app/screens/main/sections/devices/device/style.scss
