+++ public/app/screens/main/sections/devices/devices.jsx
