+++ public/app/screens/main/sections/devices/room/index.js
