+++ public/app/screens/main/sections/devices/room/room.jsx
@@ -1,7 +1,7 @@
 'use strict';
 
 var React = require('react');
-var Device = require('screens/main/devices/device/index');
+var Device = require('../device');
 
 require('./style.scss');
