+++ public/app/screens/main/sections/devices/room/style.scss
