+++ public/app/screens/main/sections/devices/style.scss
